refact: useSelectionHighlight 훅 구현
